refactor(logic): split setup ownership and capital tile scan for #2288
Extract provinceNeighboursFromTopology to game_setup_topology.dart.
Move GP landmass budget/packing helpers into game_setup_ownership part file.
Move capital tile candidate scan helpers into capital_choice part file.

Brings game_setup_ownership.dart and capital_choice.dart under 700 lines
with no behavior change. Refs #2288.
