+++ packages/colonizethis_logic/lib/src/setup/game_setup_topology.dart
@@ -1,0 +1,20 @@
+import 'package:colonizethis_data/colonizethis_data.dart';
+
+/// Province–province adjacency from topology edges (P–P links only).
+Map<String, Set<String>> provinceNeighboursFromTopology(MapTopology topology) {
+  final provinces = {
+    for (final n in topology.nodes)
+      if (n.type == TopologyNodeType.province) n.id,
+  };
+  final neighbours = <String, Set<String>>{
+    for (final id in provinces) id: <String>{},
+  };
+  for (final edge in topology.edges) {
+    final a = edge.id1;
+    final b = edge.id2;
+    if (!provinces.contains(a) || !provinces.contains(b)) continue;
+    neighbours[a]!.add(b);
+    neighbours[b]!.add(a);
+  }
+  return neighbours;
+}
